PMM-7145 Move function from models to management.
